requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-OAuthlib==0.9.6
pyotp==2.9.0
proxmoxer==2.0.1
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest==7.4.3
flake8==6.1.0
black==23.10.1
